ai_integrations/requirements.txt
# Multi-Platform AI Assistant Requirements

# Core Framework Dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
asyncio-mqtt==0.13.0

# Flask for legacy platform support
flask==3.0.0
flask-cors==4.0.0
flask-ask==0.9.8

# Google Cloud and Assistant
google-cloud-dialogflow==2.24.1
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0

# Microsoft Bot Framework
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-ai==4.15.0
azure-cognitiveservices-speech==1.34.0

# Amazon Alexa
ask-sdk-core==1.18.0
ask-sdk-model==1.82.0

# AI and NLP Libraries
openai==1.3.7
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Vietnamese Language Processing
underthesea==6.7.0
pyvi==0.1.1
vncorenlp==1.0.3

# Data Processing
pandas==2.1.3
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Database and Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.0

# Messaging and Communication
twilio==8.10.3
python-telegram-bot==20.7
pyzalo==1.0.0
sendgrid==6.10.0

# Authentication and Security
cryptography==41.0.8
pyjwt==2.8.0
passlib==1.7.4

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production Deployment
gunicorn==21.2.0
supervisor==4.2.5
nginx==1.25.3